refactor: decompose dispatch.rs into dispatch/ module directory
Split the 2,321-line dispatch.rs into 6 focused sub-modules
(disposition, runtime_mode, context, capability, chain, tests) with
a mod.rs that re-exports all public items. Zero behavior changes —
all pub/pub(crate) visibility preserved, all 638 tests pass.
